move base-types.ts
